client/pc/requirements.txt
PyQt6
qt_material
websockets
maturin
nuitka
peewee
pillow